Forcing Reset to happen when Bookmark is new.
Removed having to manually set the state.
Based on the documentation of React: https://react.dev/learn/preserving-and-resetting-state

This has a downside, when if a state should be kept in between exercises this needs to be handled at the Exercise.js level otherwise this is state is lost when a new exercise is given. Otherwise, it seems to work.

Removed the _resetReactStates()