Redifined defineExerciseSession
It wasn't clear how batch counts and remaining exercises were being used, and when utilizing a higher number of exercises the logic would result in undefined exercises.

Now calculateExerciseSequence takes a list of exercise types and iterates it from first to last, and adding it as long as there are enough bookmarks to fit that exercise. If there are too many bookmarks, it will start from 0 onwards, re-using exercises until the total number of bookmarks is used.

Tested in both the Exercise tab and in the Article 'Practice Words'. They both work. In Practice words, a session will be the entire words that were translated.

NUMBER_OF_BOOKMARKS_TO_PRACTICE  was changed to 10 as before this was the number of bookmarks used.