Created a docker image + container to build extensions
Created an image that links a volume between docker and the container. This container runs the npm scripts so that the scripts can be written to be run independently of the OS, and instead focus on running within the container.

At the end of the compose build commands a zip is created with the built binaries.