Improved feedback when typing
We now check if the user has typed the word in the wrong language in the case where we ask them to write in the learning language. This is to help in cases like TranslateWhatYouHear.

Made the hint when the user needs to change 1 letter to be considered a hint for the api rather than wrong.

Fixed a bug where the input should be accepted once the user had 1 distance to the correct answer if they type in L1.

Co-Authored-By: merleschoen <147062323+merleschoen@users.noreply.github.com>