Support Endless scrolling
This only works when combined with the changes introduced in https://github.com/zeeguu/api/pull/154

Move the functions to get the scroll location to a utility function as they might be useful in other places. I decide to use absolute pixels to avoid expanding the articles to much and loading the articles despite not having scrolled to the end.

Otherwise the users will see the loading animation at the bottom of the articles indefinitely.

Question: Should we limit the total articles a user can retrieve? At the moment I can see that they can keep going, I scrolled for a while and I got to 639 articles, it will still work but maybe it is a lot of information to keep storing all the articles?