Fixed error of articleListRef.current
Since this method was being added in an async call, we would sometimes create the event after we would leave the home page. Which leads to it never being cleared.

This implementation always creates the event, and deletes it whenever it's left from the method. The event itself has a guard in case the user attempts to scroll when no more articles are available which avoids making unnecessary calls to the API.