Fix Edit Form Sending Toasty when No edit was made
I found out that the condition to test if anything had been edited from the edit form was incorrect.

It was trying to access an undefined property 'fitForStudy' instead of the correct 'fit_for_study'.

I have moved this to a constant to be easier to call rather than having multiple places where it's re-written.