Allow users to decided if they always want to see the total number of exercises
- Initially, we simply would add a notification dot to tell users that there are exercises to review. This saved the API some calls.
- However, performance doesn't change that much and some users might want to know how many bookmarks they have accumulated. The higher the worse it becomes.

To Note, users with too many bookmarks to review might be a bit overwhelmed as they could have more than 200 and the number won't change. As a suggestion, if we decide to move with this change we should clear the bookmarks for any users which have been inactive or have more than 50 bookmarks scheduled (which is now our maximum). This should avoid situations where the user sees an extreme amount of bookmarks to review.