Update the redirect function, and GoToButton.
Update the redirect function from utils/routing to let developers decide whether links should be opened in the same or a new window.
Apply the redirect function to the onClick event handler of the GoToButton instead of using a href attribute of its link wrapper.
The button's link wrapper is to be deleted in further commits.