Update useActivityTimer.js
- Update the useIdleTimer props.

I believe that the reason we see very long sessions might have to do with the fact that the idleTimer doesn't stop once a user has left the webpage. It's hard to identify the culprit as there are multiple interactions at play, but I have seen this be alt+tabing and seeing that the counter keeps ticking. To address this I propose the following:

- We specify exactly which elements should "activate" the timer. By default, mouseMove + vibilityChange are included, but I have noted that visibilityChange is only turned off if the window is hidden completely from view. This could be the reason why sometimes the timer doesn't activate the onIdle.
- I also remove onMouseMove, as I think a timeout of 30 seconds between a mouse clicking/scroll is enough of a bandwidth to allow users to interact with the content
- throttle was unsued. That function is called onAction which we do not use. Instead I enabled eventsThrottle, to save CPU time from checking all the events from every 200ms to every 500ms

I am hoping these changes might fix the issue of very long time sessions, but I still think they will be a bit difficult to fully capture.