Do not show Open link for videos
- This is a Hotfix rather than an actual update.

The bookmarks currently only take the article id and source id, which aren't enough to open a video. We either need to pass the video_id with the bookmark, or provide an endpoint which provides the type + id to the frontend so we can render the appropriate component given a SourceID.

For now, it will display a "(Video)" so the user knows it came from a video.